Break up extensions duck
